requirements.txt
# Framework Web
Flask==3.0.3
Flask-Cors==4.0.1
Werkzeug==3.0.3
gunicorn==21.2.0

# Base de données
pymongo==4.7.2

# Gestion des PDF
PyMuPDF==1.24.4
pypdf==4.2.0

# IA et Traitement du texte
sentence-transformers==2.7.0
torch==2.3.1
transformers==4.41.1
numpy==1.26.4

# Sécurité et Authentication
PyJWT==2.8.0
python-dotenv==1.0.1

# Requêtes HTTP
requests==2.32.2

# Utilitaires
tqdm==4.66.4
pillow==10.3.0

# Logging et monitoring
psutil==5.9.8